+++ GeosCore/hco_interface_gc_mod.F90
@@ -4892,7 +4892,7 @@ CONTAINS
            eflx(I,J,NA) = eflx(I,J,NA) + tmpFlx
 
            ! Compute column emission fluxes for satellite diagnostics
-           IF ( State_Diag%Archive_SatDiagn ) THEN
+           IF ( State_Diag%Archive_SatDiagnColEmis ) THEN
               tmpFlx = 0.0_fp
               DO L = 1, State_Grid%NZ
                  CALL GetHcoValEmis( Input_Opt, State_Grid, NA,    I,        &
@@ -4900,11 +4900,7 @@ CONTAINS
                  IF ( .NOT. found ) EXIT
                  tmpFlx = tmpFlx + emis
               ENDDO
-
-              !$OMP CRITICAL
-              ! Try placing this reduction into a critical block
               colEflx(I,J,NA) = colEflx(I,J,NA) + tmpFlx
-              !$OMP END CRITICAL
            ENDIF
 
         ENDIF
@@ -4956,8 +4952,9 @@ CONTAINS
     !$OMP PARALLEL DO                                                        &
     !$OMP DEFAULT( SHARED )                                                  &
     !$OMP PRIVATE( I,       J,            N                                 )&
-    !$OMP PRIVATE( thisSpc, dep                                             )&
-    !$OMP PRIVATE( ND,      fracNoHg0Dep, zeroHg0Dep, Hg_cat                )
+    !$OMP PRIVATE( thisSpc, dep,          S                                 )&
+    !$OMP PRIVATE( ND,      fracNoHg0Dep, zeroHg0Dep                        )&
+    !$OMP COLLAPSE( 2                                                       )
     DO J = 1, State_Grid%NY
     DO I = 1, State_Grid%NX
 
@@ -5038,47 +5035,28 @@ CONTAINS
        !=====================================================================
        State_Chm%SurfaceFlux(I,J,:) = eflx(I,J,:) - dflx(I,J,:) ! kg/m2/s
 
-! Temporarily disable this code, it causes segmentation faults
-! Investigate later on -- Bob Yantosca, 23 May 2022
-!       !=====================================================================
-!       ! Defining Satellite Diagnostics
-!       !=====================================================================
-!
-!       ! Define emission satellite diagnostics
-!       IF ( State_Diag%Archive_SatDiagnColEmis ) THEN
-!
-!         ! Point to mapping obj specific to this diagnostic
-!         mapData => State_Diag%Map_SatDiagnColEmis
-!
-!          DO S = 1, mapData%nSlots
-!             N = mapData%slot2id(S)
-!             !Print *, 'JDS N = ', N
-!             State_Diag%SatDiagnColEmis(:,:,S) = colEflx(:,:,N)             
-!          ENDDO
-!
-!        ! Free pointer
-!        mapData => NULL()
-!
-!       ENDIF
-!
-!       ! N.B. SatDiagnSurfFlux contains within it the underlying eflx
-!       ! variable as opposed to colEflx.
-!       ! Thus, taking SatDiagnSurfFlux - SatDiagnColEmis will not
-!       1  necessarily equal the dry deposition flux (dflx)
-!       IF ( State_Diag%Archive_SatDiagnSurfFlux ) THEN
-!
-!         ! Point to mapping obj specific to this diagnostic
-!         mapData => State_Diag%Map_SatDiagnSurfFlux
-!
-!          DO S = 1, mapData%nSlots
-!             N = mapData%slot2id(S)
-!             State_Diag%SatDiagnSurfFlux(:,:,S) = State_Chm%SurfaceFlux(:,:,N)
-!          ENDDO
-!
-!        ! Free pointer
-!        mapData => NULL()
-!
-!       ENDIF
+       !=====================================================================
+       ! Defining Satellite Diagnostics
+       !=====================================================================
+
+       ! Define emission satellite diagnostics
+       IF ( State_Diag%Archive_SatDiagnColEmis ) THEN
+          DO S = 1, State_Diag%Map_SatDiagnColEmis%nSlots
+             N = State_Diag%Map_SatDiagnColEmis%slot2id(S)
+             State_Diag%SatDiagnColEmis(:,:,S) = colEflx(:,:,N)
+          ENDDO
+       ENDIF
+
+       ! N.B. SatDiagnSurfFlux contains within it the underlying eflx
+       ! variable as opposed to colEflx.
+       ! Thus, taking SatDiagnSurfFlux - SatDiagnColEmis will not
+       ! necessarily equal the dry deposition flux (dflx)
+       IF ( State_Diag%Archive_SatDiagnSurfFlux ) THEN
+          DO S = 1, State_Diag%Map_SatDiagnSurfFlux%nSlots
+             N = State_Diag%Map_SatDiagnSurfFlux%slot2id(S)
+             State_Diag%SatDiagnSurfFlux(:,:,S) = State_Chm%SurfaceFlux(:,:,N)
+          ENDDO
+       ENDIF
 
        !=====================================================================
        ! Archive Hg deposition for surface reservoirs (cdh, 08/28/09)