+++ CHANGELOG.md
@@ -4,6 +4,12 @@ This file documents all notable changes to the GEOS-Chem repository since versio
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [14.1.0] - TBD
+### Added
+- New satellite diagnostic (SatDiagn) collection, to archive several
+  fields within a user-defined local-time interval.  CAVEAT: For now,
+  only one local-time interval is permitted.
+
 ## [14.0.1] - 2022-10-31
 ### Fixed
 - Corrected units in metadata for State_Met%AirNumDen and State_Met%PHIS